Separate out ts and js prisma generated clients
