+++ pages/api/does-phonenumber-exist.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next"
 
-import { prisma } from "../../prisma/generated"
+import { prisma } from "../../prisma/generated/ts"
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
   const doesPhoneNumberExist = await prisma.$exists.user({
