+++ prisma/seeds.ts
@@ -3,7 +3,7 @@
 // 2. Go to your terminal make sure you are in `loan-application-platform/prisma/` directory
 // 3. run `env-cmd -f ../.config/dev.env ts-node seeds.ts`
 
-import { prisma } from "./generated"
+import { prisma } from "./generated/ts"
 import flushDB from "./flushDB"
 
 const seedDatabase = async () => {
